[Refactor]: 컴포넌트화 및 책 추가 및 수정 페이지 적용
