+++ src/components/organisms/BookInfo/BookImg.jsx
@@ -1,0 +1,10 @@
+export function BookImg() {
+	return (
+		<div className="bookImg mr-[20px]">
+			<img
+				className="self-stretch grow shrink basis-0"
+				src="https://via.placeholder.com/191x241"
+			/>
+		</div>
+	);
+}
